# docs/api-contract.openapi.yaml
openapi: "3.1.0"
info:
  title: "Club Informatique de l’IST : contrat d’API"
  version: "1.0.0"
  description: |
    Contrat de référence entre le frontend et le backend.

    Conventions : dates en ISO 8601 UTC ; pagination uniforme (« content », « page », « size », « totalElements », « totalPages ») ;
    erreurs au format RFC 9457 avec un code stable et des messages en français ; identifiants numériques stables.
    Le jeton d’accès est transmis par l’en-tête « Authorization » ; le jeton de rafraîchissement par un cookie HttpOnly.
    L’extension « x-roles » liste les rôles autorisés ; un Formateur et un Responsable du Club héritent des droits du Membre, un Super Admin de ceux de l’Administrateur.
    Les droits sont appliqués par le serveur, par rôle et par propriété de la ressource.
servers:
  - url: /api/v1
...
security:
  - jetonAcces: []
paths:
...
  "/admin/users":
    get:
      tags: [Administration]
      summary: Comptes non supprimés.
      operationId: getAdminUsers
      x-roles: [ADMIN]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/search"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/Role"
        - name: statut
          in: query
          schema:
            $ref: "#/components/schemas/StatutCompte"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageCompte"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/users/{id}":
    get:
      tags: [Administration]
      summary: Compte.
      operationId: getAdminUsersId
      x-roles: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Compte"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
    put:
      tags: [Administration]
      summary: Modifier l’identité d’un compte.
      operationId: putAdminUsersId
      x-roles: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/CompteMiseAJour"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Compte"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
  "/admin/users/{id}/roles":
    put:
      tags: [Administration]
      summary: Attribuer les rôles ; refuse d’élever un compte au-dessus du sien et de modifier ses propres rôles.
      operationId: putAdminUsersIdRoles
      x-roles: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/RolesSaisie"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Compte"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
        "409":
          $ref: "#/components/responses/Erreur409"
  "/admin/users/{id}/status":
    patch:
      tags: [Administration]
      summary: Suspendre ou réactiver un compte ; refusé sur soi-même et sur le dernier Super Admin.
      operationId: patchAdminUsersIdStatus
      x-roles: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/StatutCompteSaisie"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Compte"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
        "409":
          $ref: "#/components/responses/Erreur409"
  "/admin/users/{id}/deverrouillage":
    post:
      tags: [Administration]
      summary: Lever le verrouillage posé après des échecs de connexion.
      operationId: postAdminUsersIdDeverrouillage
      x-roles: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Opération réalisée, sans contenu.
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
  "/admin/users/invitations":
    post:
      tags: [Administration]
      summary: "Inviter une personne : elle choisit son mot de passe par un lien à usage unique."
      operationId: postAdminUsersInvitations
      x-roles: [ADMIN]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Invitation"
      responses:
        "204":
          description: Opération réalisée, sans contenu.
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "409":
          $ref: "#/components/responses/Erreur409"
  "/admin/roles":
    get:
      tags: [Administration]
      summary: Rôles et permissions effectives.
      operationId: getAdminRoles
      x-roles: [ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefinitionRole"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/admin/permissions":
    get:
      tags: [Administration]
      summary: Permissions.
      operationId: getAdminPermissions
      x-roles: [ADMIN]
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
...
components:
  securitySchemes:
    jetonAcces:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Numéro de page, à partir de zéro.
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
      description: Taille de page.
    sort:
      name: sort
      in: query
      schema:
        type: string
      description: "Tri : « champ,asc » ou « champ,desc »."
    search:
      name: search
      in: query
      schema:
        type: string
      description: Recherche plein texte ; facultative.
...
  responses:
    Erreur400:
      description: "Requête invalide : erreurs de champ dans « errors »."
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur401:
      description: Session absente ou expirée, ou identifiants refusés.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur403:
      description: Droits insuffisants (rôle ou propriété de la ressource).
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur404:
      description: Ressource introuvable.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur409:
      description: Conflit avec l’état actuel des données.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
...
  schemas:
    Probleme:
      type: object
      description: Erreur au format RFC 9457. Le champ « code » est stable ; « detail » est rédigé en français.
      required: [title, status, code]
      properties:
        type:
          type: [string, "null"]
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int64
        detail:
          type: [string, "null"]
        instance:
          type: [string, "null"]
        code:
          type: string
        errors:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/ErreurChamp"
    ErreurChamp:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    Role:
      type: string
      enum: [MEMBRE, FORMATEUR, RESPONSABLE_CLUB, ADMIN, SUPER_ADMIN, DSI]
    StatutCompte:
      type: string
      enum: [ACTIF, INACTIF, SUSPENDU, EN_ATTENTE_ACTIVATION]
...
    Compte:
      type: object
      required: [id, nom, prenom, email, statut, roles, verrouille, createdAt]
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        filiere:
          type: [string, "null"]
        statut:
          $ref: "#/components/schemas/StatutCompte"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        verrouille:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CompteMiseAJour:
      type: object
      required: [nom, prenom]
      properties:
        nom:
          type: string
        prenom:
          type: string
        filiere:
          type: [string, "null"]
    RolesSaisie:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    StatutCompteSaisie:
      type: object
      required: [statut]
      properties:
        statut:
          $ref: "#/components/schemas/StatutCompte"
    Invitation:
      type: object
      required: [nom, prenom, email, role]
      properties:
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
    DefinitionRole:
      type: object
      required: [nom, permissions]
      properties:
        nom:
          $ref: "#/components/schemas/Role"
        description:
          type: [string, "null"]
        permissions:
          type: array
          items:
            type: string
    Permission:
      type: object
      required: [nom, description]
      properties:
        nom:
          type: string
        description:
          type: string
...
    PageCompte:
      type: object
      description: Page de résultats (pagination uniforme).
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Compte"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...